move telemetry setup to the service
