+++ telemetry/wrapper.go
@@ -3,6 +3,7 @@ package telemetry
 import (
 	"time"
 
+	"cosmossdk.io/server/v2/telemetry"
 	"github.com/hashicorp/go-metrics"
 )
 
@@ -27,7 +28,7 @@ func ModuleMeasureSince(module string, start time.Time, keys ...string) {
 	metrics.MeasureSinceWithLabels(
 		keys,
 		start.UTC(),
-		append([]metrics.Label{NewLabel(MetricLabelNameModule, module)}, globalLabels...),
+		append([]metrics.Label{NewLabel(MetricLabelNameModule, module)}, telemetry.GlobalLabels...),
 	)
 }
 
@@ -38,36 +39,36 @@ func ModuleSetGauge(module string, val float32, keys ...string) {
 	metrics.SetGaugeWithLabels(
 		keys,
 		val,
-		append([]metrics.Label{NewLabel(MetricLabelNameModule, module)}, globalLabels...),
+		append([]metrics.Label{NewLabel(MetricLabelNameModule, module)}, telemetry.GlobalLabels...),
 	)
 }
 
 // IncrCounter provides a wrapper functionality for emitting a counter metric with
 // global labels (if any).
 func IncrCounter(val float32, keys ...string) {
-	metrics.IncrCounterWithLabels(keys, val, globalLabels)
+	metrics.IncrCounterWithLabels(keys, val, telemetry.GlobalLabels)
 }
 
 // IncrCounterWithLabels provides a wrapper functionality for emitting a counter
 // metric with global labels (if any) along with the provided labels.
 func IncrCounterWithLabels(keys []string, val float32, labels []metrics.Label) {
-	metrics.IncrCounterWithLabels(keys, val, append(labels, globalLabels...))
+	metrics.IncrCounterWithLabels(keys, val, append(labels, telemetry.GlobalLabels...))
 }
 
 // SetGauge provides a wrapper functionality for emitting a gauge metric with
 // global labels (if any).
 func SetGauge(val float32, keys ...string) {
-	metrics.SetGaugeWithLabels(keys, val, globalLabels)
+	metrics.SetGaugeWithLabels(keys, val, telemetry.GlobalLabels)
 }
 
 // SetGaugeWithLabels provides a wrapper functionality for emitting a gauge
 // metric with global labels (if any) along with the provided labels.
 func SetGaugeWithLabels(keys []string, val float32, labels []metrics.Label) {
-	metrics.SetGaugeWithLabels(keys, val, append(labels, globalLabels...))
+	metrics.SetGaugeWithLabels(keys, val, append(labels, telemetry.GlobalLabels...))
 }
 
 // MeasureSince provides a wrapper functionality for emitting a a time measure
 // metric with global labels (if any).
 func MeasureSince(start time.Time, keys ...string) {
-	metrics.MeasureSinceWithLabels(keys, start.UTC(), globalLabels)
+	metrics.MeasureSinceWithLabels(keys, start.UTC(), telemetry.GlobalLabels)
 }
